requirements.txt
h5py
Orange3>=3.20
numpy
scipy
matplotlib
silx
